% Generated by roxygen2 (4.0.1): do not edit by hand
\name{word2vec}
\alias{word2vec}
\title{Train a model by word2vec.}
\usage{
word2vec(train_file, output_file)
}
\arguments{
\item{train_file}{Path of the train file.}

\item{output_file}{Path of the output file.}
}
\value{
A word2vec object.
}
\description{
Train a model by word2vec.
}
\details{
The word2vec tool takes a text corpus as input and produces the
word vectors as output. It first constructs a vocabulary from the
training text data and then learns vector representation of words.
The resulting word vector file can be used as features in many
natural language processing and machine learning applications.
}
\examples{
\dontrun{
word2vec(system.file("examples", "rfaq.txt", package = "tmcn.word2vec"))
}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
\references{
\url{https://code.google.com/p/word2vec/}
}

